# internal/types/get_stream_response.go
// Code generated by go-swagger; DO NOT EDIT.

package types

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
)

// GetStreamResponse get stream response
// Example: channel1
//
// swagger:model getStreamResponse
type GetStreamResponse string

// Validate validates this get stream response
func (m GetStreamResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this get stream response based on context it is used
func (m GetStreamResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}